requirements.txt
#attrs==22.2.0
#backcall==0.2.0
#certifi==2022.12.7
#colorama==0.4.3
daiquiri==3.2.1
#decorator==4.4.2
#exceptiongroup==1.1.0
#iniconfig==2.0.0
#ipykernel==5.3.4
#ipython==7.18.1
#ipython-genutils==0.2.0
#jedi==0.17.2
#jupyter-client==6.1.7
#jupyter-core==4.6.3
#packaging==23.0
#parso==0.7.1
#pickleshare==0.7.5
#pip==23.0.1
#pluggy==1.0.0
#prompt-toolkit==3.0.7
#Pygments==2.7.1
pytest==7.2.2
#python-json-logger==2.0.7
PyYAML==6.0
#setuptools==65.6.3
#tomli==2.0.1
#tornado==6.0.4
#traitlets==5.0.4
#wcwidth==0.2.5
#wheel==0.38.4
#wincertstore==0.2
